call_analysis_pipeline/pyproject.toml
[project]
name = "call-analysis-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.5",
    "langchain>=1.0.8",
    "langchain-openai>=1.0.3",
    "openai>=2.8.1",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "streamlit>=1.51.0",
    "tiktoken>=0.12.0",
]
